Fixing up validation. Testing removal of trimming non printing characters.
